+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade Swaggy C to 6.2.2
 - Upgrade OpenAPI Generator to 7.24.0
 
+### Fixed
+- Fix ECONNRESET when a pooled conn is used in test JS
+
 ## 6.0.0 - 2026-01-24
 ### Added
 - Add application/javascript response content